src/third_party/blink/web_tests/external/wpt/svg/interact/parsing/pointer-events-valid-expected.txt

This is a testharness.js-based test.
PASS e.style['pointer-events'] = "auto" should set the property value
PASS e.style['pointer-events'] = "bounding-box" should set the property value
FAIL e.style['pointer-events'] = "visiblePainted" should set the property value assert_equals: serialization should be canonical expected "visiblePainted" but got "visiblepainted"
FAIL e.style['pointer-events'] = "visibleFill" should set the property value assert_equals: serialization should be canonical expected "visibleFill" but got "visiblefill"
FAIL e.style['pointer-events'] = "visibleStroke" should set the property value assert_equals: serialization should be canonical expected "visibleStroke" but got "visiblestroke"
PASS e.style['pointer-events'] = "visible" should set the property value
PASS e.style['pointer-events'] = "painted" should set the property value
PASS e.style['pointer-events'] = "fill" should set the property value
PASS e.style['pointer-events'] = "stroke" should set the property value
PASS e.style['pointer-events'] = "all" should set the property value
PASS e.style['pointer-events'] = "none" should set the property value
Harness: the test ran to completion.
